requirements.txt
# ActionInbox Dependencies
flask>=2.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
python-dateutil>=2.8.0
requests>=2.28.0
discord.py>=2.3.0
python-dotenv>=1.0.0